CREATE2 Factory Deployment Dapp
This PR creates a small Dapp on a static HTML page for deploying the
CREATE2 factory contract. It is very spartan, and uses almost no
libraries (beyond `@noble/{curves|hashes}` for the required
cryptography and fits nicely into <20kb compressed!

**Unfortunately**, all wallets I tested (including MetaMask and Brave
Wallet) does not currently support EIP-7702 type `0x4` transaction
with the `authorizationList` field, meaning that we cannot execute the
bootstrapping transaction on the deployer account. In fact, MetaMask
produces the following error:

```
External EIP-7702 transactions are not supported
```

What is worse is that Brave Wallet with just happily ignore the
`authorizationList` field and will happily execute the transaction
without any authorizations.

That being said, we have a work around by creating a burner account
that is used to actually execute the deployment.