Update title (#6)
I am a little hesitant to call it "Universal", this word is overused and
imo has a negative sentiment for a lot of devs (as there have been many
"universal" proposals). For me the benefit of this factory is the
deployment method, which makes it "permissionless" and "cross-chain".
Therefore I would propose to go back to the original title
("Permissionless Create2 Factory") and if we want we could add
cross-chain to it.

Another alternative: Permissionless, cross-chain contract factory
deployment