Propagate Revert Data
This PR changes the permissionless CREATE2 factory to propagate revert
data (like [EIP-7997](https://github.com/ethereum/EIPs/pull/10092))
instead of just reverting with empty data.

Note that we no longer have the init code fiting into 32-byte :'(, but
the runtime code still does :). That being said, the usability
improvement of propagating revert data is worth the extra code size
IMO. Note that deployment gas costs _still remain the same_, as the
extra code is just in the `revert` branch for contract deployments.

I also chose to keep `calldatasize() - 32` overflow behaviour as it is
now. That is, it will revert with OOG (from trying to copy ~2**256
bytes of calldata to memory) instead of handling the revert manually.
The rationale here is:

1. The call reverts one way or the other, so we don't have bad
   state (different from the suggestion in EIP-7997 which uses a
   saturating subtraction, where incorrect deployments will silently
   fail)
2. Additional gas costs would be paid on **every** create2 deployment
   to handle the conditional jump (basically, the "if calldatasize() <
   32" check), for arguably very little benefit: it will only save gas
   for incorrect invocations, which shouldn't be used anywhere anyway.