CREATE2 Factory Deployment Dapp in React
This is a re-write of #2 but using React + Wagmi + Mantine. It
definitely looks much better, is reactive, and _should_ be more
stable (as we use standard tools instead of handrolling everything
ourselves).

I am very much a React noob - so I may have made some mistakes along
the way.

The bundle size is around 10x that of the vanilla JS app, but still a
manageable 200kB gzipped.