Increase Grace Period Post-Transaction Confirmation
This PR increases the grace period after confirming a transaction.
During testing, @tschubotz found that for some chains, the `Deploy`
button would become enabled again, which could cause some issues as
the user would try to deploy the factory a second time (which would
result in an error).

Although, I can't reproduce it locally, my intuition is that the stale
empty bytecode for the permissionless CREATE2 factory that is used to
determin the state of the deployment wizard takes longer to refresh in
some situations. The solution is to increase the "grace period" on the
transaction button (so the button's "confirmed" state lingers for
longer) and remove the "stale time" setting on one of our queries
related to the current app state - making it so requeries always
refetch data instead of reusing a, potentially stale, cache.